Initial metadata joiner - combines corpus data and pg_catalog
Reads in the metadata generated by the SPGC and combines it
with the pg_catalog.csv that is posted by Project Gutenberg.

Note that the info in pg_catalog.csv is omitted from the
metadata file created by SPGC and could be added to that
spreadsheet upon creation, if desired.

After reading the data in, drops anything that is not a text.
Note that several things that don't have type specified are currently
included, until they are examined.

Adds tools to compare columns, to ensure pg_catalog and metadata.csv
match