GutenbergDataLoader supports skip_start_and_end
Moves skip_start_and_end to the GutenbergDataLoader.
Uses the same method of doing this as when it was outside the class,
which involves reading the entire text in and then stripping it.
This was required before, but can be done more efficiently now that it
is all in one data loading class.  Will be updated in one of the next few
commits.